pyproject.toml
[project]
name = "arcfusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "bs4>=0.0.2",
    "fastapi[standard]>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.1",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "typer>=0.16.0",
]
